split up parts of page into elements
