+++ components/protectedRoute.tsx
@@ -2,8 +2,9 @@ import { useEffect } from "react";
 import { useAuth } from "./authProvider";
 import { useRouter } from "next/router";
 export default function ProtectedRoute({children}:any){
-    const {isLoading, isAuthenticated} = useAuth();
+    const {isLoading, isAuthenticated, user} = useAuth();
     const router = useRouter();
+
     if(isLoading){
         return(
             <p>Loading...</p>
