+++ pages/connections/signIn.tsx
@@ -19,5 +19,6 @@ export default async function SignInWithGoogle() {
     });
 
 
+    alert(data)
     
   }
